requirements.txt
aerich==0.7.1
aiosqlite==0.17.0
anyio==3.6.2
APScheduler==3.10.0
asyncpg==0.27.0
atlastk==0.13.2
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dictdiffer==0.9.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
lxml==4.9.2
psycopg2==2.9.5
pydantic==1.10.5
pypika-tortoise==0.1.6
python-dotenv==0.21.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==1.4.46
starlette==0.25.0
tomlkit==0.11.6
tortoise-orm==0.19.3
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
